requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
lxml==4.5.2
multitasking==0.0.9
numpy==1.19.1
pandas==1.1.1
pandas-datareader==0.9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shyaml==0.6.1
six==1.15.0
tabulate==0.8.7
urllib3==1.25.10
yfinance==0.1.54
